DABOM-215 refactor: UsagePersistService 분리
